Rename functions and create lib folder
