+++ cmd/reset.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/GoToolSharing/htb-cli/config"
 	"github.com/GoToolSharing/htb-cli/utils"
+	"github.com/GoToolSharing/htb-cli/utils/webhooks"
 	"github.com/spf13/cobra"
 )
 
@@ -66,7 +67,7 @@ var resetCmd = &cobra.Command{
 			log.Fatalf("Error: %v", err)
 		}
 		if config.GlobalConf["Discord"] != "False" {
-			err := utils.SendDiscordWebhook("[RESET] - " + output)
+			err := webhooks.SendToDiscord("[RESET] - " + output)
 			if err != nil {
 				fmt.Println(err)
 				return
