+++ cmd/sherlocks.go
@@ -108,7 +108,7 @@ var sherlocksCmd = &cobra.Command{
 	Short: "Displays active sherlocks and next sherlocks to be released",
 	Run: func(cmd *cobra.Command, args []string) {
 		if sherlockNameParam != "" {
-			sherlockID, err := sherlocks.SearchSherlockIDByName(proxyParam, sherlockNameParam, batchParam)
+			sherlockID, err := sherlocks.SearchIDByName(proxyParam, sherlockNameParam, batchParam)
 			if err != nil {
 				fmt.Println(err)
 				return
@@ -116,7 +116,7 @@ var sherlocksCmd = &cobra.Command{
 			log.Println("SherlockID :", sherlockID)
 
 			if sherlockTaskID != 0 {
-				err := sherlocks.GetSherlockTaskByID(proxyParam, sherlockID, sherlockTaskID)
+				err := sherlocks.GetTaskByID(proxyParam, sherlockID, sherlockTaskID)
 				if err != nil {
 					fmt.Println(err)
 					return
@@ -124,14 +124,14 @@ var sherlocksCmd = &cobra.Command{
 				return
 			}
 
-			err = sherlocks.GetSherlockGeneralInformations(proxyParam, sherlockID, sherlockDownloadPath)
+			err = sherlocks.GetGeneralInformations(proxyParam, sherlockID, sherlockDownloadPath)
 
 			if err != nil {
 				fmt.Println(err)
 				return
 			}
 
-			data, err := sherlocks.GetSherlockTasks(proxyParam, sherlockID)
+			data, err := sherlocks.GetTasks(proxyParam, sherlockID)
 			if err != nil {
 				fmt.Println(err)
 				return
